pyproject.toml
[project]
name = "ai-minds-hackathon"
version = "0.1.0"
description = "AI Minds Hackathon"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "qdrant-client",
    "neo4j",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
